EID-641 & EID-642: Remove optional RequestedAttributes
- For the time being, we are only requesting the mandatory minimal
  attributes from member states.
- Move EidasAuthnRequestFromHubToAuthnRequestTransformer tests to the
  right place.

Author: @vixus0
